+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add support to render template using `host` or `$(uname -n)` as the value of `overlay show --render`. #623
 - Added command line parameters for credentials of a container registry
 - Add flag `--build` to `wwctl container copy`. #1378
+- Add `wwctl clean` to remove OCI cache and overlays from deleted nodes
 
 ### Changed
 
@@ -38,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove NodeInfo (in-memory-only) data structure, consolidating onto NodeConf. #916
 - Replace `defaults.conf` with settings on the default profile. #917
 - Switched from yaml.v2 to yaml.v3 #1462
+- Make OCIBlobCache a seperate path and point it to `/var/cache` #1459
 
 
 ### Removed